list-view: add API to retrieve artifacts from a list view
First endpoints returning a Flux typed with Event<? ReactivityEntity> are introduced by this commit to:
- retrieve a mocked organization with a single list view
- subscribe to this mocked organization and receive associated view with its corresponding artifacts
- retrieve some additional artifacts matching a specific period of time

The view selects the 100 newest artifacts when the user subscribes to the organization. A mocked index.html shows how this web service can be called and then retrieve older artifacts based on the last received events.

The endpoints are currently not suspending a SSE connection because Reactor currently broadcast any event to all connections, which is not what we want for this use case. It's plan in a future release of Reactor to provide such capability.

Flux are configured to fallback to an Event<Error> publisher in case of exception or timeout. This mechanism needs to be improved with retry capabilities, including replica usage for couchbase, that meet circuit-breaking capabilities.

There's still a random issue (under investigation) where no artifacts is returned by N1QL query in couchbase the first time a web service is called. Also, log4j2 is raising an exception when called by Reactor through SLF4J and that's why it's not present in this commit.