versioning: filter document versions
This commit filters the documents according to the application version. Versions older than the application version are accepted since backward compatibility is expected. If the version is not a snapshot, then any document tagged as snapshot will be ignored. To achieve this easily at query level, the semver value is converted to an integer representation and the snapshot flag is isolated in a dedicated field. For now, each part of the semver value (MAJOR, MINOR and PATCH) is expected to be on 1 digit (1.0.0 is ok but not 1.1.10). There is also no consideration about broken compatibility between two major versions (document in 1.x.x will be returned even if the application version is 2.x.x). Those rules can be reviewed in the future if needed.