deployment: push front module to CloudFoundry
Statics are packaged inside a dist directory which is generated when the build has been successfully executed.
Then, the Staticfile buildback is executed to upload the directory content to CloudFoundry.
The deployed content is currently not optimized for production (no minification, caching, etc).
Also note that all bower components (JS and HTML files) are copied without a module import resolution process.
Currently the deployment should not really work as API URL can't be configured while is it provisionned dynamically by CloudFoudry.
A future commit will fix it.