config: exposes the broadcaster domain in a /api-domain.json endpoint
Web application can now resolve the backend domain in a /api-domain.json endpoint.
When deploying to CloudFoundry, gulp is used to generate this file before upload starts.
The domain is retrieved from the manifest.yml where route is declared.
For development, gulp handles this endpoint and specifies the localhost:3000/api domain.