security: add anonymous authentication support
Core spring security and spring session components are leveraged to build a custom support that is servlet-api agnostic, allowing to build an ExpiringSession and an Authentication objects on top of non servlet containers like netty.

The project currently uses the spring-webflux WebFilter to enable session creation by capturing all requests with a cookie. However, spring-webflux WebSession component is currently not really configurable and priority is given to the spring-session ExpiringSession that can be backed by a distributed system support like Hazelcast. Using directly the WebSession will be the preferred approach as soon as it's possible to configure it.

Security layer relies on the Authentication and UserDetailsService components of spring-security. However, a custom WebFilter is used to reproduce a similar mechanism on top of spring-webflux by extracting an Authentication object backed by a ThreadLocal. This allows to bind bean lifecycle to the reactivity session scope and autowire the current user authentication to them. The security layer currently generates an anonymous user with a session automatically refreshed when it expires. The infrastructure is designed to handle pre-authenticated authentication requests, where it is assumed that the principal has already been authenticated by an external system.