restdocs: adds spring-rest-docs
This commit introduces the spring-rest-docs support.
All EventController endpoinds are documented and generated snippets are pushed to a "doc-snippets" branch that can be used in the reactivity-doc project.
Currently unit tests are mixing MockMvc with webflux module, which works even if a more adapted support for webflux tests has to be considered.